Fix resolution of eager macro contents
